pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "codesense"
version = "0.4.0"
description = "AI-powered semantic code search with intelligent parsing for Python projects"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Artem", email = "stasiuk.dev@gmail.com"}
]
maintainers = [
    {name = "Artem", email = "stasiuk.dev@gmail.com"}
]
keywords = [
    "codesense",
    "semantic-search",
    "code-search",
    "embeddings",
    "faiss",
    "sentence-transformers",
    "ai",
    "machine-learning",
    "nlp",
    "developer-tools"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Code Generators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
]
requires-python = ">=3.9"
dependencies = [
    "sentence-transformers>=2.0.0",
    "faiss-cpu>=1.7.0",
    "click>=8.0.0",
    "tqdm>=4.65.0",
    "numpy>=1.21.0",
    "torch>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/steliarix/codesense"
Documentation = "https://github.com/steliarix/codesense#readme"
Repository = "https://github.com/steliarix/codesense"
Issues = "https://github.com/steliarix/codesense/issues"
Changelog = "https://github.com/steliarix/codesense/blob/main/CHANGELOG.md"

[project.scripts]
codesense = "codesense.cli:cli"

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312', 'py313']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = []

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --cov=codesense --cov-report=term-missing"
